requirements.txt
streamlit
langchain
langchain-community
langchain-groq
python-dotenv
sentence-transformers
chromadb
scikit-learn
pypdf
pandas

